Move primekeys out of primenodes.cpp to their own file.
