+++ src/makefile.linux-mingw
@@ -70,6 +70,7 @@ OBJS= \
     obj/noui.o \
     obj/kernel.o \
     obj/primenodes.o \
+    obj/primekeys.o \
     obj/scrapesdb.o
 
 all: paycoind.exe
