+++ src/makefile.mingw
@@ -123,6 +123,7 @@ OBJS= \
     obj/noui.o \
     obj/kernel.o \
     obj/primenodes.o \
+    obj/primekeys.o \
     obj/scrapesdb.o
 
 all: paycoind.exe
